feat(router): middleware support
Created struct Application, which is our own Abstraction layer for stuff like middleware and routing, with tview.Application embedded.
It does feel strange that my first basic router lib, is for a non networking project, especially a TUI that honestly doesn't need it, but I thought it would be fun.